requirements.txt
google-auth-oauthlib==1.2.0
google-api-python-client==2.121.0
pytest==8.1.1
tqdm==4.67.1